lesson245: adding coaches to vuex
