+++ src/store/getters.js
@@ -1,1 +1,5 @@
-export default {}
+export default {
+  showId(state) {
+    return state.userId
+  }
+}
