+++ src/store/mutations.js
@@ -1,1 +1,5 @@
-export default {}
+export default {
+  addAuthUserId(state, payload) {
+    state.userId = payload;
+  }
+}
